+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Recommended Lists for Refresh Shelf.
+
 ## [0.4.0] - 2020-09-03
 
 ### Added